requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
Automat==25.4.16
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
constantly==23.10.4
cryptography==45.0.5
decorator==5.2.1
distlib==0.3.9
ecdsa==0.19.1
fastapi==0.115.14
filelock==3.18.0
frozenlist==1.7.0
graphql-core==3.2.6
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.12
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==0.18.3
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
strawberry-graphql==0.278.0
tornado==6.5.1
trio==0.30.0
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uv==0.7.21
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
